:truck: Move: model->entity 이동
